requirements.txt
autopep8==1.4.4
coverage==4.5.3
Django==2.2.3
django-bootstrap4==0.0.8
django-model-utils==3.2.0
django-nose==1.4.6
djangorestframework==3.9.4
nose==1.3.7
Pillow==6.1.0
psycopg2==2.8.3
pycodestyle==2.5.0
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
virtualenv==16.6.2
